% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rr_cluster.R
\name{add_geneset_columns}
\alias{add_geneset_columns}
\title{Add geneset-specific columns to cluster dataframe}
\usage{
add_geneset_columns(cluster_df, merged_gs, gs_names)
}
\arguments{
\item{cluster_df}{Cluster dataframe}

\item{merged_gs}{Original merged geneset}

\item{gs_names}{Geneset names}
}
\value{
Updated cluster dataframe
}
\description{
Add geneset-specific columns to cluster dataframe
}
\keyword{internal}
