% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_tab.R
\name{build_cluster_params}
\alias{build_cluster_params}
\title{Build clustering parameters from UI inputs}
\usage{
build_cluster_params(input, method)
}
\arguments{
\item{input}{Shiny input object}

\item{method}{Selected clustering method}
}
\value{
Named list of parameters
}
\description{
Build clustering parameters from UI inputs
}
\keyword{internal}
