% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rr_cluster.R
\name{cluster_david}
\alias{cluster_david}
\title{DAVID functional clustering}
\usage{
cluster_david(merged_gs, params, gs_names)
}
\arguments{
\item{merged_gs}{Merged geneset dataframe}

\item{params}{List with: similarity_threshold, initial_group_membership,
final_group_membership, multiple_linkage_threshold}

\item{gs_names}{Geneset names for output}
}
\value{
Clustering result list
}
\description{
Uses richCluster::david_cluster for DAVID-style functional clustering
with multiple linkage thresholds.
}
\keyword{internal}
