% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_tab.R
\name{clusterTabServer}
\alias{clusterTabServer}
\title{Cluster Tab Server}
\usage{
clusterTabServer(
  id,
  u_degnames,
  u_degdfs,
  u_rrnames,
  u_rrdfs,
  u_big_rrdf,
  u_clusnames,
  u_clusdfs,
  u_big_clusdf,
  u_cluslists,
  clus_intermed
)
}
\arguments{
\item{id}{Module namespace ID}

\item{u_degnames}{Reactive values for DEG names}

\item{u_degdfs}{Reactive values for DEG dataframes}

\item{u_rrnames}{Reactive values for enrichment result names}

\item{u_rrdfs}{Reactive values for enrichment result dataframes}

\item{u_big_rrdf}{Reactive values for combined enrichment data}

\item{u_clusnames}{Reactive values for cluster result names}

\item{u_clusdfs}{Reactive values for cluster result dataframes}

\item{u_big_clusdf}{Reactive values for combined cluster data}

\item{u_cluslists}{Reactive values for cluster list details}

\item{clus_intermed}{Reactive values for intermediate clustering results}
}
\description{
Cluster Tab Server
}
