% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_tab.R
\name{clusterTabUI}
\alias{clusterTabUI}
\title{Cluster Tab UI}
\usage{
clusterTabUI(id, tabName)
}
\arguments{
\item{id}{Module namespace ID}

\item{tabName}{Tab identifier for shinydashboard}
}
\value{
A tabItem containing the clustering UI
}
\description{
Cluster Tab UI
}
