% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rr_cluster.R
\name{get_cluster_list_richR}
\alias{get_cluster_list_richR}
\title{Build cluster list from richR clustering result}
\usage{
get_cluster_list_richR(clustered_gs, merged_gs, gs_names)
}
\arguments{
\item{clustered_gs}{Output from richR::richCluster}

\item{merged_gs}{Original merged geneset}

\item{gs_names}{Geneset names}
}
\value{
Dataframe with cluster assignments and term details
}
\description{
Build cluster list from richR clustering result
}
\keyword{internal}
