% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rr_cluster.R
\name{normalize_geneset}
\alias{normalize_geneset}
\title{Normalize enrichment dataframe}
\usage{
normalize_geneset(df)
}
\description{
Coerces column types and fills required columns to avoid downstream errors.
}
\keyword{internal}
