% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{_PACKAGE}
\name{richStudio}
\alias{richStudio}
\title{richStudio: A package for FGSEA and clustering}
\description{
richStudio: A package for FGSEA and clustering
}
\section{richStudio functions}{

The richStudio functions ...
}

