% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_handling.R
\name{rm_file_clusdf}
\alias{rm_file_clusdf}
\title{Remove Cluster Result from File Table}
\usage{
rm_file_clusdf(df, rm_vec)
}
\arguments{
\item{df}{Existing file tracking dataframe}

\item{rm_vec}{Dataframe containing cluster results to remove (must have 'name' column)}
}
\value{
Updated file tracking dataframe with specified cluster results removed
  Returns NULL if df is atomic
}
\description{
Removes specified cluster results from the master file tracking table.
}
\examples{
\dontrun{
file_table <- rm_file_clusdf(file_table, file_table[1, ])
}

}
