% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_handling.R
\name{rm_file_degdf}
\alias{rm_file_degdf}
\title{Remove DEG Set from File Table}
\usage{
rm_file_degdf(df, rm_vec)
}
\arguments{
\item{df}{Existing file tracking dataframe}

\item{rm_vec}{Dataframe or vector containing DEG sets to remove (must have 'name' column)}
}
\value{
Updated file tracking dataframe with specified DEG sets removed
  Returns NULL if df is atomic
}
\description{
Removes specified DEG sets from the master file tracking table.
}
\examples{
\dontrun{
file_table <- rm_file_degdf(file_table, file_table[1, ])
}

}
