% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_tab.R
\name{saveTabServer}
\alias{saveTabServer}
\title{Save Tab Server}
\usage{
saveTabServer(
  id,
  u_degnames,
  u_degdfs,
  u_rrnames,
  u_rrdfs,
  u_clusnames,
  u_clusdfs,
  u_cluslists
)
}
\arguments{
\item{id}{Module namespace ID}

\item{u_degnames}{Reactive values for DEG names}

\item{u_degdfs}{Reactive values for DEG dataframes}

\item{u_rrnames}{Reactive values for enrichment result names}

\item{u_rrdfs}{Reactive values for enrichment result dataframes}

\item{u_clusnames}{Reactive values for cluster result names}

\item{u_clusdfs}{Reactive values for cluster result dataframes}

\item{u_cluslists}{Reactive values for cluster list details}
}
\description{
Save Tab Server
}
