% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_tab.R
\name{saveTabUI}
\alias{saveTabUI}
\title{Save Tab UI}
\usage{
saveTabUI(id, tabName)
}
\arguments{
\item{id}{Module namespace ID}

\item{tabName}{Tab identifier for shinydashboard}
}
\value{
A tabItem containing the save/load UI
}
\description{
Save Tab UI
}
