% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_tab.R
\name{updateTabUI}
\alias{updateTabUI}
\title{Update Tab UI}
\usage{
updateTabUI(id, tabName)
}
\arguments{
\item{id}{Module namespace ID}

\item{tabName}{Tab identifier for shinydashboard}
}
\value{
A tabItem containing the update UI
}
\description{
Update Tab UI
}
